[AUTOCODEX] refactor
